ver 0.0.11
- participantsNumber in event calcuates automatically
- MyUser + authoredEvents attribute
- Deleteing events removes event from author authoredEvents
- MyUser deleted correction (deleteing user delete all authoredEvents of user).

+++ src/main/java/com/example/eventorestapi/service/UserService.java
@@ -1,6 +1,8 @@
 package com.example.eventorestapi.service;
 
+import com.example.eventorestapi.exceptions.NotExistException;
 import com.example.eventorestapi.exceptions.UserAlreadyExistsException;
+import com.example.eventorestapi.models.Event;
 import com.example.eventorestapi.models.MyUser;
 import com.example.eventorestapi.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +10,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Iterator;
 import java.util.Optional;
 
 @Service
@@ -15,6 +18,9 @@ public class UserService {
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private EventService eventService;
 
     @Autowired
     private BCryptPasswordEncoder bCryptPasswordEncoder;
@@ -36,8 +42,25 @@ public class UserService {
         return userRepository.findByEmail(email);
     }
 
+    @Transactional
     public void removeUserByEmail(String email) {
+        Optional<MyUser> userOpt = userRepository.findByEmail(email);
+        if (userOpt.isEmpty()) {
+            throw new NotExistException("User", "email");
+        }
+        MyUser user = userOpt.get();
+        deleteAllEventsOfUser(user);
         userRepository.deleteByEmail(email);
+    }
+    @Transactional
+    public void deleteAllEventsOfUser(MyUser user){
+        //Iterator To Avoid ConcurrentModificationException
+        Iterator<Event> authoredEventsIterator = user.getAuthoredEvents().iterator();
+        while (authoredEventsIterator.hasNext()) {
+            Event event = authoredEventsIterator.next();
+            authoredEventsIterator.remove();
+            eventService.deleteEvent(event);
+        }
     }
 
 }
